+++ src/paid-cookie-consent/audit-data-provider.js
@@ -97,7 +97,9 @@ function calculateBounceGapLoss(bounceGapData, log) {
 
   const groupedByTrafficType = {};
   for (const row of bounceGapData) {
-    if (!groupedByTrafficType[row.trfType]) groupedByTrafficType[row.trfType] = {};
+    if (!groupedByTrafficType[row.trfType]) {
+      groupedByTrafficType[row.trfType] = {};
+    }
     groupedByTrafficType[row.trfType][row.consent] = {
       pageViews: row.pageViews,
       bounceRate: row.bounceRate,