+++ test/audits/cwv-trends-audit/handler.test.js
@@ -10,7 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-/* eslint-env mocha */
 import { expect } from 'chai';
 
 describe('CWV Trends Audit Handler', function () {